[Packages] Load packages async

+++ Zebra/ZBSettings.h
@@ -152,8 +152,8 @@ NS_ASSUME_NONNULL_BEGIN
 + (BOOL)alwaysInstallLatest;
 + (void)setAlwaysInstallLatest:(BOOL)alwaysInstallLatest;
 
-+ (int16_t)role;
-+ (void)setRole:(int16_t)role;
++ (uint8_t)role;
++ (void)setRole:(uint8_t)role;
 
 + (NSArray *)ignoredUpdates;
 + (BOOL)areUpdatesIgnoredForPackageIdentifier:(NSString *)identifier;
